Add new and create action for plc
